Feedback -  Deleted unused code
